+++ src/assets/AssetDefinitions.ts
@@ -25,7 +25,7 @@ export interface AssetDefinitions {
   texture: { resource: Texture; config: { source: string; mimeType?: string; textureOptions?: Partial<TextureOptions> } };
   sound: {
     resource: Sound;
-    config: { source: string; playbackOptions?: Partial<PlaybackOptions>; poolSize?: number; sprites?: Readonly<Record<string, AudioSpriteClip>> };
+    config: { source: string; playbackOptions?: Partial<PlaybackOptions>; poolSize?: number; sprites?: Readonly<Record<string, AudioSpriteClip>> | string };
   };
   music: {
     resource: AudioStream;